+++ src/main/java/org/springboot/trendmartecommerceplatform/Product/ProductController.java
@@ -2,9 +2,9 @@ package org.springboot.trendmartecommerceplatform.Product;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
-import io.swagger.v3.oas.annotations.security.SecurityScheme;
-import lombok.Getter;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -32,17 +32,22 @@ public class ProductController {
 
     @PostMapping
     @Operation(summary = "Inserting a new product")
+    @PreAuthorize("hasRole('ADMIN')")
     public Product addProduct(@RequestBody Dto dto) {
         return productService.addProduct(dto);
     }
-    @PatchMapping("/update/{id}")
-    @Operation(summary = "Update products")
-    public Product updateProduct(@RequestBody Dto dto, @PathVariable long id) {
+
+    @PreAuthorize("hasRole('ADMIN')")
+    @PatchMapping("/admin/edit/{id}")
+    @Operation(summary = "Admin Updates/edits products")
+    public Product EditProduct(@Valid @RequestBody Dto dto, @PathVariable long id) {
         return productService.updateProduct(id, dto);
     }
 
+
     @DeleteMapping
     @Operation(summary = "Delete product by id")
+    @PreAuthorize("hasRole('ADMIN')")
     public void deleteProduct(@PathVariable long id) {
         productService.deleteProduct(id);
     }
@@ -50,3 +55,41 @@ public class ProductController {
 
 
 }
+//
+//package org.springboot.trendmartecommerceplatform.Product;
+//
+//import io.swagger.v3.oas.annotations.Operation;
+//import jakarta.validation.Valid;
+//import lombok.AllArgsConstructor;
+//import org.springframework.security.access.prepost.PreAuthorize;
+//import org.springframework.web.bind.annotation.*;
+//
+//@RestController
+//@RequestMapping("/api/admin/products")
+//@AllArgsConstructor
+//public class AdminProductController {
+//    private final ProductService productService;
+//
+//    @PreAuthorize("hasRole('ADMIN')")
+//    @PostMapping
+//    @Operation(summary = "Admin adds products")
+//    public Product addProduct(@Valid @RequestBody Dto dto) {
+//        return productService.addProduct(dto);
+//    }
+//
+//    @PreAuthorize("hasRole('ADMIN')")
+//    @PatchMapping("/edit/{id}")
+//    @Operation(summary = "Admin Updates/edits products")
+//    public Product EditProduct(@Valid @RequestBody Dto dto, @PathVariable long id) {
+//        return productService.updateProduct(id, dto);
+//    }
+//
+//    @PreAuthorize("hasRole('ADMIN')")
+//    @PatchMapping("/delete/{id}")
+//    @Operation(summary = "Admin delets products")
+//    public void DeleteProduct(@Valid  @PathVariable long id) {
+//        productService.deleteProduct(id);
+//    }
+//
+//}
+
