--- src/main/java/org/springboot/trendmartecommerceplatform/config/MethodSecurityConfig.java
@@ -1,9 +1,0 @@
-package org.springboot.trendmartecommerceplatform.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
-
-@Configuration
-@EnableMethodSecurity
-public class MethodSecurityConfig {
-}
